Updated background tiles to prepare for double resolution grid
Since every tile now has 4 squares on it that the bike can move to, I
needed to draw out every possible configuration a tile could be in the
bike.chr file. Each of the four squares could either be open, crossed by
player 1, or crossed by player 2, this leaves a total of 81 unique tiles
that had to be created. (3^4)