Perfected the movement system
The game will now remember if a directional button is pressed and held.
The game will take any second direction as a priority over the held one
and set the second direction as the next place on to move.  When the
second button is released the bike will switch back to the first held
direction if it is still held.  If the first held button is released,
the second held button will take its place as the first held.

To recap,
No held button: the bike moves in the same direction without change.
One held button: the bike will turn (if allowed) to move in the
direction pressed and will keep going that way even after release.
Two held buttons: the bike will turn (if allowed) to move in the
direction of the second button to be pressed.  If the second button is
released, the bike goes back in the first held direction (if still
held).  If the first button is released, the second direction will take
its place as the first direction and the bike will continue moving that
way even if all buttons are released.