Now allows for updating the background, and refreshing the grid after a round ends
Every time the bike moves to a new square the Picture Processing Unit
(PPU) can be updated to replace an old tile with a new one.  This can be
seen in this current build as the grid will start out clean at the start
of a round, then the number one will appear in the top left corner of
the grid as the bike begins to move, and is wiped away when the round
ends.